Update roadmap due to required refactoring and tech debt payback
The Contour maintainers have recently realised that some of our big
ticket items (like external authentication and upgrading from Envoy's
xDS v2 to v3 APIs) are very difficult with our current code structure,
with a couple of years of tech debt.

So we are taking some time to pay down some tech debt and refactor,
to build a base from which we can build all the extensibility that you,
our users are asking for.

We've been conservative with the roadmap updates here, these timeframes
allow for us to spend a bit more time now to save a lot of time later.

Signed-off-by: Nick Young <ynick@vmware.com>